+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Common Changelog](https://common-changelog.org/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Validate inputs with exceptions instead of `assert` (which is stripped under `python -O`), read variable attributes only when they will be displayed, and truncate variable names to the configured `--column-widths` instead of a hardcoded width ([#349](https://github.com/nasa/ncompare/issues/349)) ([**@danielfromearth**](https://github.com/danielfromearth))
+
 ## [1.14.0] - 2025-12-30
 
 _Updates base Python version from 3.9 to 3.11._